+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * New: Add `composer.json` for PHPUnit Polyfill dependency
     * Hardening: Prevent direct web access to `bootstrap.php`
     * Change: In bootstrap, store path to plugin directory in a constant
+    * New: Add tests for `c2c_parentless_categories()`
     * New: Add tests for function invocation filters
     * Change: Tweak some inline comment formatting